Fix opacity of popover arrows.
Previously, the percentage of the LESS `fadein` function was being
used directly as the opacity, which resulted in an opacity of `0.05`
rather than `0.25` as the LESS version generates.

We can use the SASS `fade_in` function to more accurately generate
the correct color using the same values; the only difference is that
SASS's `fade_in` expects a decimal value, rather than the percentage
that LESS accepts.